Activate daRkStudio

Activate daRkStudio

activate(index_file = NULL, backup = TRUE)

Arguments

index_file
character: Path to RStudio's index.htm. Useful for times when the default installation method cannot successfully locate the file.

backup
logical: TRUE or FALSE. Copies the default index.htm file to index.htm.pre-ds. Defaults to TRUE.

daRkStudio modifies index.htm, a file used by RStudio to construct it's DOM (Document Object Model).

The only change to index.htm the inclusion of a <link> element near the end of the file, which tells RStudio to load darkstudio.css.

daRkStudio creates a directory, "darkstudio", inside the www folder, which is found at /Applications/RStudio.app/Contents/Resources/www on macOS, and C:\Program Files\RStudio\www on Windows.

activate() will create a backup of index.htm at www/darkstudio/index.htm.pre-ds. www/darkstudio is also where you will find darkstudio.css, which is the bread and butter of this package.

On Windows, you will likely need Administrator Privileges if you've installed RStudio to the default location, C:\Program Files\RStudio.

Value

Returns TRUE if the operation is successful.

Examples

if (FALSE) {
# Default:
activate()

# macOS:
index_htm <- "/Applications/RStudio.app/Contents/Resources/www/index.htm"
activate(backup = TRUE, index_file = index_htm)

# Windows:
index_htm <- "C:/Program Files/RStudio/www/index.htm"
activate(backup = TRUE, index_file = index_htm)
}